updated gen script to error if no grammar file provided

+++ gen.sh
@@ -1,4 +1,11 @@
 #!/bin/bash
+if [ -z "$1" ]
+    then
+        echo "No grammar file supplied"
+        echo "Suggested usage:"
+        echo "./gen.sh popl.g4"
+        exit 1
+fi
 ANTLR="antlr-4.11.1-complete.jar"
 rm -rf parserlibs
 mkdir parserlibs
